refactor(transcription): move LocalModelConfig type to local services folder
Relocate LocalModelConfig type from shared types to local transcription services directory for better code organization. Updates all import statements to reference the new location.
