+++ apps/whispering/src/lib/services/transcription/local/types.ts
